Change : Update Libswscale to latest Libav revision. Fix : Crash internal audio decoder on some PCM stream. Fix : x64 build crash at play video. Code cleanup.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/branches/ffmpegupdate@3761 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/transform/MPCVideoDec/ffmpeg/ffImgfmt.h
@@ -59,7 +59,7 @@ enum ffdshow_colorspaces {
 #define FF_CSP_FLAGS_YUV_JPEG   (1U<<27)
 
 #include <stddef.h>
-typedef ptrdiff_t stride_t;
+typedef int stride_t;
 
 //==================================== xvid4 =====================================
